pyproject.toml
[project]
name = "homework"
version = "0.1.0"
description = ""
authors = [
    {name = "dvodnenko",email = "thekovalskii@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-simple-captcha (>=0.6.2,<0.7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
